Make bot register url configurable
With the bot now in staging, we need to have a second development
version of the bot inside discord which will have a different invite
url, so we can't use the old hardcoded one any longer. This was also a
chance to streamline the html templating and get rid of all the static
files.