Add profiling and supress log feature
To investigate performance, we now have a feature flag that allows to
make a flamegraph of the whole program runtim.

We now should be able to control the max level of emitted logs at
compile time. It might be nicer in the future to also have a feature
flag around the log level enum, to only allow values up to the max
level.

Sled seems to somehow still escape our loggin supression....