Add check cli command
To measure performance progress, we needed a way to test our business
logic. Since it is difficult to benchmark binaries directly in stable
rust and also the problems of micro benchmarking, we now have an own
cli endpoint, that exposes the most expensive operations of reputation
fetching for all gates on a server.

This might maybe also come handy when deubgging some role problems
later.