Prevent accidental spillage of secrets
We don't want to accidentally reveal the discord token or encryption key
in logs, so wrapping them in a secret string type prevents that and also
handles some improvements in dropping the memory. The other possible
spillage can happen when the config sources are shown in the
`config show` cli subcommand. There we know filter out the critical
strings.