Fix round reputation to zero bug
Because of the generalization of our gating conditions also the
reputation conversion logic got changed. In this case we converted an
f64 into a u64 which was zero for values smaller than one. We now do the
calculations correctly by scaling up the f64 with a factor that is very
large but small enough to not overflow with a 100 percent reputation
threshold, which is the maximum the user can enter.