Improve tracing instrumentation
We want to be able to follow the control flow through the application,
with the right context in each situation but without being too verbose.
For this reason we suppress the logging of external crates for log
levels higher than trace. Only http access logs are allowed in addition
to our own logic. We also try to be as useful with the tracing
spans as possible and include enough but not to much data.