Streamline config management
The behavior on how config values got merged was a little off and
sometimes broken. We now do what is promised and use Default, Env, Cli
and File configs in witch correct precedences.

Default values are now handled correctly and we are also able to print
how the config is composed.

Some commands need to bypass config parsing to also work when the config
is not complete.