Make colony clients generic
To make the code of our gates unit testable, we want to have the ability
to create mock instances of the clients that don't make real http requests
or requests towards the blockchain.

One problem that arose in this context was, that  deserialization with typetag can't handle
generic structs, so we can't just give our gates structs a generic client field.
Instead we used a static client instance which is a little harder to set up.